// apps/web/lib/white-label/schema/signup-link-form.schema.ts
import { z } from 'zod';

import { CreditsSchema } from './credits-form.schema';

export const SignupLinkFormSchema = z.object({
  signup: z.object({
    name: z.string().min(2).max(50),
    code: z.string().min(5).max(30),
    max_usage: z.coerce.number().int().min(0).optional(),
    expires_at: z.coerce.date().optional(),
  }),
  permissions: CreditsSchema,
});